setting up comments for future save
we'll need to work out how we're doing both saves or one or the other.  If it's not to heavy on the game we could save and pack the blob and then we could save local binary file to the harddrive.   OR we could just utf8 encode it if that's not to much to do twice